% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QADF.R
\name{QADF}
\alias{QADF}
\title{Quantile ADF test}
\usage{
QADF(variable, taus, dq_max)
}
\arguments{
\item{variable}{The time-series, which is tested for stationarity}

\item{taus}{Vector of quantiles in (0, 1)}

\item{dq_max}{Maximal value of lag, which is used for lag selectiion}
}
\value{
Dataframe with quantiles and corresponding coefficient before $alpha$,
       marked with ***, **, * corresponding, to $\alpha$ \geq 1 P-value (10%, 5%, 1%)
}
\description{
Calculates auantile ADF test statistic based on (Koenker, Xiao, 2004)
$Q(y_t \mid I_t, tau) = \mu_1(\tau) + \mu_2(\tau) + \alpha y_{t - 1} + \sum_{j = 1}^{dq} \alpha_{j} \Delta y_{t - j}$
$H_0$: Non-stationary -> $H_0: \alpha = 1$
}
