% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QECM.R
\name{QECM}
\alias{QECM}
\title{Quantile Error Correction model with step selection}
\usage{
QECM(p, q1, q2, p_lag, q1_lag, q2_lag, tau_list)
}
\arguments{
\item{p}{Dependent variable}

\item{q1}{Independent variable 1}

\item{q2}{Inpendendt variable 2}

\item{p_lag}{Maximum lag of p}

\item{q1_lag}{Maximum lag of q1}

\item{q2_lag}{Maximum lag of q2}

\item{tau_list}{List of quantiles for fit}
}
\value{
Dataframe with fitted coefficients
}
\description{
Fits QECM and selects lags
}
