% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QGranger.R
\name{QGranger}
\alias{QGranger}
\title{Quantile Granger Test}
\usage{
QGranger(X, Y, qar_order_max, taus)
}
\arguments{
\item{X}{Independent variable}

\item{Y}{Dependent variable}

\item{qar_order_max}{Order of Y autoregression}

\item{taus}{List of quantiles}
}
\value{
P-values of F_test
}
\description{
Performs a Granger Test with F-stat
}
