% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{QQplot}
\alias{QQplot}
\title{Quantile-in-Quantile Plot}
\usage{
QQplot(QQresult)
}
\arguments{
\item{QQresult}{QQresult matrix}
}
\value{
Surface Plot with tau_X and tau_Y on x and y axis and corresponding beta
}
\description{
Plots Quantile-in-Quantile parameters againts quantiles of $X$ and $Y$
}
