% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ.R
\name{QQRegression}
\alias{QQRegression}
\title{Quantile-in-Quantile Regression}
\usage{
QQRegression(Y, X, tau_Y, tau_X, h = 0.05)
}
\arguments{
\item{Y}{Dependent variable}

\item{X}{Independent variable}

\item{tau_Y}{List of qunatiles of Y}

\item{tau_X}{List of quantiles of X}

\item{h}{Kernel width}
}
\value{
QQ Result 3d matrix with beta(i, j) s. t. Q(Y, tau(i)) = beta(i, j) Q(X, T(j)) + eps(i, j)$
}
\description{
Fit Quantile-in-Quantile Regression with Gaussian kernel
}
