% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{ts_plot_ly}
\alias{ts_plot_ly}
\title{Time-series plots with linear trend}
\usage{
ts_plot_ly(y, dates)
}
\arguments{
\item{y}{series}

\item{dates}{date-index}
}
\value{
plotly line plot
}
\description{
Plots plotly graph for y in time with linear trend
}
